chore: Add ArtistSong model and association
This commit adds the ArtistSong model and establishes an association between the Artist and Song models. The ArtistSong model includes foreign keys 'artistId' and 'songId' that reference the 'id' columns of the Artist and Song models respectively. This association allows an artist to be associated with multiple songs and a song to be associated with multiple artists.